Fix nil nodes rendered as the Go placeholder <nil>
Per the YAML 1.2 specification, a null value may be represented as an
empty key. Rendering `<nil>` wouldn't be valid under any circumstances.

This commit fixes the previously introduced test case, by ensuring a <nil> node
is rendered as an empty string instead, which in turn correctly represents the
YAML specification for valid null output (which may be an empty string).